pyproject.toml
[tool.poetry]
name = "hourth_challenge"
version = "0.1.0"
description = ""
authors = ["Victor Rafael <victorsls@outlook.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.1.1"
djangorestframework = "^3.13.1"
drf-yasg = "^1.21.3"
requests = "^2.28.1"
pandas = "^1.5.0"

[tool.poetry.dev-dependencies]
autopep8 = "^1.7.0"
black = "^22.8.0"
autoflake = "^1.6.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
